refactor: get my comments to my page controller
