+++ src/main/java/com/codeit/moim/service/mypage/impl/MyPageServiceImpl.java
@@ -3,21 +3,21 @@ package com.codeit.moim.service.mypage.impl;
 import com.codeit.moim.common.exception.auth.PasswordInvlaidException;
 import com.codeit.moim.common.exception.auth.UserNotFoundException;
 import com.codeit.moim.common.exception.payload.ErrorStatus;
-import com.codeit.moim.domain.Contact;
-import com.codeit.moim.domain.Skill;
-import com.codeit.moim.domain.User;
-import com.codeit.moim.domain.UserSkill;
-import com.codeit.moim.repository.ContactRepository;
-import com.codeit.moim.repository.SkillRepository;
-import com.codeit.moim.repository.UserRepository;
-import com.codeit.moim.repository.UserSkillRepository;
+import com.codeit.moim.domain.*;
+import com.codeit.moim.repository.*;
 import com.codeit.moim.service.mypage.MyPageService;
 import com.codeit.moim.service.storage.StorageService;
 import com.codeit.moim.service.user.impl.UserServiceImpl;
+import com.codeit.moim.web.dto.request.comment.ReadMyCommentRequest;
 import com.codeit.moim.web.dto.request.mypage.*;
+import com.codeit.moim.web.dto.response.comment.ReadMyCommentResponse;
 import com.codeit.moim.web.dto.response.member.CreateMemberResponse;
 import com.codeit.moim.web.dto.response.mypage.*;
+import com.codeit.moim.web.dto.response.slice.CustomSlice;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Slice;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -35,6 +35,7 @@ public class MyPageServiceImpl implements MyPageService {
     private final ContactRepository contactRepository;
     private final UserSkillRepository userSkillRepository;
     private final SkillRepository skillRepository;
+    private final CommentRepository commentRepository;
     private final UserServiceImpl userServiceImpl;
     private final PasswordEncoder passwordEncoder;
 
@@ -119,6 +120,30 @@ public class MyPageServiceImpl implements MyPageService {
         userRepository.save(user);
 
         return new UpdatePasswordResponse(userId);
+    }
+
+    @Override
+    public Slice<ReadMyCommentResponse> getMyComments(int userId, ReadMyCommentRequest request) {
+        int pageSize = request.size();
+        Pageable pageable = PageRequest.of(0, pageSize);
+
+        Slice<Comment> comments;
+        if(Objects.isNull(request.lastCommentId()) || request.lastCommentId() <=0 ) {
+            comments = commentRepository.findByUser_userIdOrderByCommentIdDesc(userId, pageable);
+        }else{
+            comments = commentRepository.findByUser_UserIdAndCommentIdLessThanOrderByCommentIdDesc(userId, request.lastCommentId(), pageable);
+        }
+
+        List<ReadMyCommentResponse> commentResponses = comments.stream()
+                .map(comment ->
+                        ReadMyCommentResponse.fromEntity(comment, comment.getMeeting())
+                ).collect(Collectors.toList());
+
+        Integer nextCursor = comments.hasNext()
+                ? comments.getContent().get(comments.getContent().size() -1).getCommentId()
+                : null;
+
+        return new CustomSlice<>(commentResponses, pageable, comments.hasNext(), nextCursor);
     }
 
     private User getUser(int userId){
